functions/pre_upload_check/requirements.txt
boto3==1.39.4
botocore==1.39.4
jmespath==1.0.1
numpy==1.26.4
python-dateutil==2.9.0.post0
s3transfer==0.13.0
six==1.17.0
urllib3==2.5.0
python-jose==3.4.0